back/requirements.txt
Flask-RESTful
PyJWT
Flask-SQLAlchemy
Flask_marshmallow
flask_restful
SQLAlchemy<1.4.0
flask_cors